+++ src/i18n/ptBR/index.ts
@@ -29,3 +29,15 @@ export const aboutTranslation = {
   fallbackContent:
     '<p>Olá! Eu me chamo <strong>João Bispo</strong>, comecei como desenvolvedor de software em 2015, tentando rotacionar um boi na minha tela com minhas primeiras linhas de código usando JavaScript.</p><p>Atualmente sou <strong>engenheiro de software</strong> na IBM. Desde o meu início, gosto muito de resolver problemas com criação de software usando conceitos de engenharia de software, eficiência de algoritmos e escolher a ferramenta certa para o trabalho. Sou do <strong>Brasil</strong>, quero conhecer outros países e cidades, mas sempre morei aqui.</p><p>Nos meus tempos livres gosto de fomentar a minha criatividade, fazendo <strong>PixelArt</strong>, escrever artigos, ver animes e jogar. Eu também gosto de malhar e <strong>milkshake de chocolate</strong>.</p>',
 };
+
+export const roadmapTranslation = {
+  title: 'Mais alto. Mais longe. Mais ágil.',
+  subtitle:
+    'Este é meu roadmap, eu irei atualizá-lo conforme eu aprender coisas novas e melhorar minhas habilidades que eu deseje compartilhar o progresso.',
+  seoTitle: 'Roadmap',
+  seoDescription:
+    'Este é meu roadmap, eu irei atualizá-lo conforme eu aprender coisas novas e melhorar minhas habilidades.',
+  titleTimeManage: 'A matriz de Eisenhower',
+  imageTimeManageDescription:
+    'A matriz de Eisenhower é uma ferramenta de gestão de tempo que ajuda a priorizar as tarefas de acordo com a importância eurgência.',
+};