+++ src/pages/blog/posts/[slug].tsx
@@ -1,6 +1,13 @@
-import { Flex, Heading, HStack, Icon, Text } from '@chakra-ui/react';
+import {
+  Flex,
+  Heading,
+  Icon,
+  IconButton,
+  Text,
+  useDisclosure,
+} from '@chakra-ui/react';
 import { GetStaticPaths, GetStaticProps, NextPage } from 'next';
-import { FiCalendar, FiUser } from 'react-icons/fi';
+import { FiCalendar, FiShare2 } from 'react-icons/fi';
 import { useRouter } from 'next/router';
 
 import { useTranslation } from '@src/hooks/useTranslation';
@@ -12,6 +19,7 @@ import { Content } from '@src/styles/poststyles';
 import { SEO } from '@src/components/SEO';
 import { Tags } from '@src/components/Tags';
 import { parseLocaleToGraphCmsLocale } from '@src/utils/parseLocale';
+import { ShareSocialMediaModal } from '@src/components/ShareSocialMedia';
 
 const PostPage: NextPage = () => {
   const { graphCmsLocale } = useTranslation();
@@ -27,6 +35,8 @@ const PostPage: NextPage = () => {
   });
 
   const { locale } = useTranslation();
+  const { isOpen, onOpen, onClose } = useDisclosure();
+  const postUrl = global.window ? window.location?.href : '';
   // TODO: ADD image={data?.post?.coverImage?.coverImagePost[0]?.coverImage?.url} post image in future
 
   return (
@@ -36,7 +46,7 @@ const PostPage: NextPage = () => {
       width="100%"
       minHeight="calc(100vh - 8rem)"
       flexDirection="column"
-      paddingTop={[4, 16]}
+      paddingTop={[4, 8]}
       paddingX={'1.75rem'}
     >
       <SEO
@@ -49,20 +59,36 @@ const PostPage: NextPage = () => {
         {data?.post?.title}
       </Heading>
       <Tags tags={data?.post?.tags} />
-      <HStack spacing="2rem" marginTop={'2rem'}>
+      <Flex gap="2rem" marginTop={'2rem'} justifyContent={'space-between'}>
         <Flex alignItems="center" gap="1rem">
-          <Icon as={FiCalendar} w={6} h={6} color="brand.secondary" />
+          <Icon
+            as={FiCalendar}
+            w={6}
+            h={6}
+            color="brand.secondary"
+            background="brand.background"
+          />
           <Text as="span" color="brand.secondary">
             {formatDate(data?.post?.date, locale)}
           </Text>
         </Flex>
         <Flex alignItems="center" gap="1rem">
-          <Icon as={FiUser} w={6} h={6} color="brand.secondary" />
-          <Text as="span" color="brand.secondary">
-            {'João Bispo'}
-          </Text>
+          <IconButton
+            aria-label="Compartilhar post"
+            icon={<FiShare2 color="brand.secondary" size={24} />}
+            color="brand.secondary"
+            variant="link"
+            onClick={onOpen}
+            transition="all 0.2s"
+            _hover={{ color: 'brand.yellow' }}
+          />
+          <ShareSocialMediaModal
+            content={postUrl}
+            isOpen={isOpen}
+            onClose={onClose}
+          />
         </Flex>
-      </HStack>
+      </Flex>
       <Content
         dangerouslySetInnerHTML={{
           __html: data?.post?.content.html ?? '',